+++ packages/core/lib/contants/index.js
@@ -1,0 +1,6 @@
+const providers = {
+    INJECTED: 'INJECTED',
+    CONNECTED: 'CONNECTED',
+};
+
+export { providers };