+++ packages/store/CHANGELOG.md
@@ -1,5 +1,14 @@
 ### versions
 
-## 2.0.0 major improvements for store and changed some variables names for reducers
+## 2.0.0
 
-## 2.1.0 major removed connection thunks and added store book reducer
+major improvements for store and changed some variables names for reducers
+
+## 2.1.0
+
+-   major removed connection thunks and added store book reducer
+-   added unit tests for reducers
+
+## 2.1.2
+
+-   fixed bug, book reducer couldn't be imported